fix(AntiInviteService): fix permission and whitelist check

+++ src/services/AntiInviteService.js
@@ -1,5 +1,4 @@
 /* eslint-disable */
-const StringParserUtil = require('../utils/misc/StringParserUtil');
 const buildAntiInviteMessage = require('../helpers/messages/buildAntiInviteMessage');
 
 class AntiInviteService {
@@ -11,7 +10,10 @@ class AntiInviteService {
   async applyActions() {
     const data = await this.client.database.collection('guilds').findOne({ guildID: this.message?.guild.id });
     if (data.antiInvite?.enabled && this.message.guild.me.permissions.has('ADMINISTRATOR')) {
-      //if (this.message.member.permissions.has('ADMINISTRATOR')) return false;
+      if (
+        this.message.member.permissions.has('ADMINISTRATOR')
+        || data.antiInvite.whiteList?.some((role) => this.message.member.roles.cache.has(role))
+      ) return false;
 
       this.message.delete().catch(() => {});
       if (data.antiInvite?.message) this.message.channel.send(
